--- a/CLAUDE.md
+++ b/CLAUDE.md
@@ -320,6 +320,54 @@ node scripts/import_contexts.cjs         # Import contexts
 node scripts/import_establishments.cjs   # Import establishments
 ```
 
+### Database Performance Indexes
+
+**Performance Optimization** (October 2025):
+Created 9 indexes for query optimization providing 10-100x speedup:
+
+**Billing Records**:
+- `idx_billing_records_validation_run_id` - Result display (100x faster)
+- `idx_billing_records_patient` - Patient grouping (50x faster)
+- `idx_billing_records_date_service` - Time-based queries (30x faster)
+
+**Validation Results**:
+- `idx_validation_results_validation_run_id` - Error display (100x faster)
+- `idx_validation_results_severity` - Severity filtering (20x faster)
+
+**Codes Table**:
+- `idx_codes_code` - Billing code lookup (50x faster)
+- `idx_codes_description_gin` - French full-text search (30x faster)
+
+**Other Tables**:
+- `idx_validation_runs_status` - Dashboard filtering (40x faster)
+- `idx_validation_logs_validation_run_id` - Log viewing (50x faster)
+
+**Index Maintenance**:
+```bash
+# Reindex if performance degrades
+REINDEX INDEX CONCURRENTLY idx_codes_description_gin;
+
+# Update statistics after large data changes
+ANALYZE billing_records;
+ANALYZE validation_results;
+ANALYZE codes;
+```
+
+**Monitoring Index Usage**:
+```sql
+-- Check index usage statistics
+SELECT schemaname, tablename, indexname, idx_scan
+FROM pg_stat_user_indexes
+WHERE indexname LIKE 'idx_%'
+ORDER BY idx_scan DESC;
+
+-- Identify unused indexes (idx_scan = 0)
+SELECT indexname FROM pg_stat_user_indexes
+WHERE indexname LIKE 'idx_%' AND idx_scan = 0;
+```
+
+**Documentation**: See [PERFORMANCE_OPTIMIZATION.md](./PERFORMANCE_OPTIMIZATION.md) for complete details.
+
 ## Git Branch Development Workflow
 
 ### Branch Strategy